Split views.py into separate files
Closes #39
