+++ job_board/views/__init__.py
@@ -1,0 +1,4 @@
+from categories import *
+from companies import *
+from jobs import *
+from misc import *
